Implemented WebSocket Endpoint to receive frames and run inference
Now it sets up an actual webserver for receiving, unlike the original main.py
The Webserver also uses Dependency injection to avoid the need for global variables. Now we have a managed state for the server connection.
Still mostly the same logic.
Also set an entrypoint script to simplify running it from uv