Implemented Dockerfile and necessary fixes for a working training run on RCP (#17)
This pull request implements the containerization and necessary changes to allow us to submit a Docker image for training on RCP.
Documentation is for now locally written, but should be pushed here for future reference.
The Dockerfile was taken from the AI Team documentation, and then modified so that I could optimize the build process. Its layers are considered in a certain order to ensure proper caching upon code changes.
In the current state, I was able to submit a working training run on a small dataset, and watch its results on WandB.

Commits involved:
* Dockerfile for RCP and training adaptations

* Setting up RCP training on A100, wandb and absolute path configs

* Dockerfile tweaks and prompt tweak

* Improvements to pyproject dependencies and docker image

* Reorganized Dockerfile, still trying more fixes

* Reorganized and optimized Dockerfile

* Improvements and fixes from bug fixing with RCP

* Tweaking optimizations

* More possible optimizations and fixes

* Trying to improve image build speeds

* Latest adaptations to the code to fix things

* Tweaks for running on RCP